Remove WeaponClass.cs and use Shared Weapon.cs
