+++ VORP-Inventory/VORP.Inventory.Client/VORP.Inventory.Client.csproj
@@ -78,7 +78,6 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Utils.cs" />
     <Compile Include="PluginManager.cs" />
-    <Compile Include="Models\WeaponClass.cs" />
   </ItemGroup>
   <ItemGroup />
   <Import Project="..\VORP.Inventory.Shared\VORP.Inventory.Shared.projitems" Label="Shared" />
