+++ VORP-Inventory/VORP.Inventory.Server/VORP.Inventory.Server.csproj
@@ -69,7 +69,6 @@
     <Compile Include="PluginManager.cs" />
     <Compile Include="Scripts\VorpPlayerInventory.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Models\WeaponClass.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
